+++ Flex121/Coordinator/EntryCoordinator.swift
@@ -1,0 +1,27 @@
+//
+//  EntryCoordinator.swift
+//  Flex121
+//
+//  Created by Rəşad Əliyev on 6/27/25.
+//
+
+import UIKit
+
+//class EntryCoordinator {
+//    
+//    private let navigationController: UINavigationController
+//    
+//    init(navigationController: UINavigationController) {
+//        self.navigationController = navigationController
+//    }
+//    
+//    func navigateToPosition() {
+//        let positionVC = PositionBuilder(coordinator: self).build()
+//        navigationController.pushViewController(positionVC, animated: true)
+//    }
+//    
+//    func navigateToGender() {
+//        let genderVC = GenderBuilder(coordinator: self).build()
+//        navigationController.pushViewController(genderVC, animated: true)
+//    }
+//}
