+++ Flex121/LoginSignUp/SignUp/Register/RegisterBuilder.swift
@@ -8,8 +8,14 @@
 import UIKit
 
 class RegisterBuilder {
+    private let coordinator: AppCoordinator
+    
+    init(coordinator: AppCoordinator) {
+        self.coordinator = coordinator
+    }
+    
     func build() -> UIViewController {
-        let viewModel = RegisterScreenViewModel()
+        let viewModel = RegisterScreenViewModel(coordinator: coordinator)
         let vc = RegisterScreenVC(viewModel: viewModel)
         return vc
     }
