+++ Flex121/LoginSignUp/SignUp/Register/RegisterScreenViewModel.swift
@@ -8,8 +8,19 @@
 import UIKit
 
 class RegisterScreenViewModel {
+    
+    private let coordinator: AppCoordinator
+    
+    init(coordinator: AppCoordinator) {
+        self.coordinator = coordinator
+    }
+    
     func goToWebsite(url: String) {
         guard let url = URL(string: url) else { return }
         UIApplication.shared.open(url)
+    }
+    
+    func goToPosition() {
+        coordinator.navigateToPosition()
     }
 }
