+++ Flex121/LoginSignUp/SignUp/Steps/Gender/GenderViewModel.swift
@@ -1,0 +1,20 @@
+//
+//  GenderViewModel.swift
+//  Flex121
+//
+//  Created by Rəşad Əliyev on 6/27/25.
+//
+
+import UIKit
+
+class GenderViewModel {
+    private let coordinator: AppCoordinator
+    
+    init(coordinator: AppCoordinator) {
+        self.coordinator = coordinator
+    }
+    
+    func goToPosition() {
+        coordinator.navigateToPosition()
+    }
+}
