+++ Flex121/LoginSignUp/SignUp/Steps/Position/PositionViewModel.swift
@@ -1,0 +1,20 @@
+//
+//  PositionViewModel.swift
+//  Flex121
+//
+//  Created by Rəşad Əliyev on 6/27/25.
+//
+
+import UIKit
+
+class PositionViewModel {
+    private let coordinator: AppCoordinator
+    
+    init(coordinator: AppCoordinator) {
+        self.coordinator = coordinator
+    }
+    
+    func goToGender() {
+        coordinator.navigateToGender()
+    }
+}
